ai-vision/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
ultralytics==8.3.0
opencv-python-headless==4.10.0.84
torch>=2.0.0
torchvision>=0.15.0
pillow==10.4.0
python-multipart==0.0.12
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
numpy<2.0
